requirements.txt
asgiref==3.6.0
dj-database-url==1.3.0
Django==4.0
djangorestframework==3.13.1
gunicorn==20.1.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2023.3
whitenoise==6.4.0
